[effector] move helper method from effector-rrt to interpolation-constraints
